+++ src/api/hooks/Cheatsheet.tsx
@@ -1,8 +1,18 @@
-import { useCallback } from 'react';
 import { Study } from 'types/Enums';
 import API from 'api/api';
+import { useInfiniteQuery } from 'react-query';
+import { getParameterByName } from 'utils';
+import { Cheatsheet, PaginationResponse, RequestResponse } from 'types/Types';
 
-export const useCheatsheet = () => {
-  const getCheatsheets = useCallback((study: Study, grade: number, filters = null) => API.getCheatsheets(study, grade, filters), []);
-  return { getCheatsheets };
+const QUERY_KEY = 'cheatsheet';
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const useCheatsheet = (study: Study, grade: number, filters?: any) => {
+  return useInfiniteQuery<PaginationResponse<Cheatsheet>, RequestResponse>(
+    [QUERY_KEY, study, grade, filters],
+    ({ pageParam = 1 }) => API.getCheatsheets(study, grade, { ...filters, page: pageParam }),
+    {
+      getNextPageParam: (lastPage) => getParameterByName('page', lastPage.next),
+    },
+  );
 };